주문 상태 변경 + 창고 반환 + 주문 취소 복구 불가

+++ myFarm/src/main/java/com/example/myFarm/inventory/InventoryMapper.java
@@ -11,7 +11,7 @@ public interface InventoryMapper {
     int initInventoryForCrop(@Param("cropId") long cropId,
                              @Param("cropName") String cropName);
     int increaseInventoryAmount(@Param("cropId") long cropId,
-                                @Param("amount") int amount);
+                                @Param("amount") long amount);
 
 
     InventoryVO getByStorId(@Param("storId") long storId);
